Branched code out into multiple sources Implemented struct for handling channel connections better (needs testing) Added basic logging of messages, will look into other events eventually
